Updating SymbolStruct printer to print things with package identifiers if not in the same package.
Found SEVERAL bugs in the NumberToken Accumulator. Still some left and I want to clean up some of the logic a bit more so it's a bit more readable and maintainable. Right now I feel like i get a bit lost on what is being checked where to ensure something can actually be parsed as a number. :(

Finally, added a bit more information to the REPL output. We not print out the Java Class's SimpleName that is returned from the reader. This should help identify what types of objects are coming out when parsing and also help determine which ones we can't print correctly. :)