Fixing the Let code generator to create the initForm stores correctly so that they can't reference each other. Also added a new LetStar code generator for when we rework the LetStar expander so we can do let* more efficiently.
Also found some bugs in the reader related to Floats and Rational Floats attempted to be created when the result should have been a symbol. Added a check for a NumberFormatException so I don't have to do ridiculous logical checks of the read attributes to determine if they are correct. This is a super minor edge case so it's ok.